refactor: 각 페이지별 init을 만들고 Main에서 호출하도록 전환
