+++ src/main.ts
@@ -1,3 +1,13 @@
+import { MovieDetail } from "./pages/movieDetail.ts";
 import { Index } from "./pages/index.ts";
+import { Search } from "./pages/search.ts";
 
-Index.init();
+const Main = {
+	init() {
+		Index.init();
+		Search.init();
+		MovieDetail.init();
+	}
+}
+
+Main.init();
