+++ src/pages/search.ts
@@ -5,6 +5,22 @@ import { MovieDetail } from "./movieDetail.ts";
 import State from "../state.ts";
 
 export const Search = {
+  init() {
+    this.setUpSearchForm();
+  },
+
+  setUpSearchForm() {
+    const searchForm = document.querySelector(".search-form");
+    searchForm?.addEventListener("submit", (event) => {
+      event.preventDefault();
+      const input = searchForm.querySelector("input");
+      if (input) {
+        const searchValue = input.value;
+        Search.showSearchMovies(searchValue);
+      }
+    });
+  },
+
   async showSearchMovies(query: string) {
     State.isLoading = true;
     Renderer.renderSkeleton(".thumbnail-list", State.requestMovieCount);
